pyproject.toml
[project]
name = "infom227-animations"
version = "0.1.0"
description = "INFOM277 Animations contains the code to generate animations for the course INFOM227."
authors = [
    { name = "Berg Lucas", email = "55436804+BergLucas@users.noreply.github.com" },
]
readme = "README.md"
license = "MIT"
keywords = ["Poetry", "INFOM227", "Manim"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
requires-python = ">=3.11,<3.13"
dependencies = [
    "manim (>=0.18.1,<0.19.0)",
    "manim-dataflow-analysis @ git+https://github.com/BergLucas/manim-dataflow-analysis.git",
    "antlr4-python3-runtime (>=4.13.2,<5.0.0)",
]

[tool.poetry]
requires-poetry = "^2.0.0"
packages = [{ include = "small", from = "src" }]

[tool.poetry.requires-plugins]
poethepoet = "^0.33.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.10"
mypy = "^1.15.0"
coverage = "^7.6.12"

[tool.poe.tasks]
test = "python -m unittest -v"
coverage = "coverage run -m unittest -v"
coverage-report = ["coverage", { cmd = "coverage report" }]
coverage-html = ["coverage", { cmd = "coverage html" }]
format = "ruff format src tests"
check = "ruff check --fix src tests"
check-nofix = "ruff check src tests"
check-unsafe = "ruff check --fix --unsafe src tests"
typecheck = "mypy src tests"
lint = ["format", "check", "typecheck"]
verify = ["lint", "test"]

[tool.coverage.run]
branch = true
source = ["src/small", "tests"]

[tool.ruff]
exclude = ["src/small/cst/small/*"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "N",   # pep8-naming
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "TCH", # flake8-type-checking
    "Q",   # flake8-quotes
    "PL",  # pylint
]

ignore = [
    "E203",
    "E501",
    "N802",
    "N818",
    "C901",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "F403",
    "F405",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
ignore_missing_imports = true
exclude = "src/small/cst/small/*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
